Revert "Merge pull request #4670 from tokyo-metropolitan-gov/development"
This reverts commit 9ee9fa1216e83f428d581bf6dae91f9dc0381f33, reversing
changes made to 373955e4c306fe160aeab8b70c5d4b3c8833b44a.
