+++ assets/locales/ja.json
@@ -465,12 +465,5 @@
   "検査実施人数（健康安全研究センターによる実施分）": "検査実施人数（健康安全研究センターによる実施分）",
   "※（4）（5）の目安となる数値は、患者数に応じた病床の確保数を参考として記載している。": "※（4）（5）の目安となる数値は、患者数に応じた病床の確保数を参考として記載している。",
   "（5）入院患者数 レベル1：1,000床 → レベル2：3,000床 → レベル3：4,000床": "（5）入院患者数 レベル1：1,000床 → レベル2：3,000床 → レベル3：4,000床",
-  "（4）重症患者数 レベル1：100床 → レベル2：300床 → レベル3：700床": "（4）重症患者数 レベル1：100床 → レベル2：300床 → レベル3：700床",
-  "{date}の全体累計": "{date}の全体累計",
-  "検査実施人数（2月14日以前分を含む累計）": "検査実施人数（2月14日以前分を含む累計）",
-  "検査実施人数 (日別)": "検査実施人数 (日別)",
-  "本グラフは、5月22日より「モニタリング指標（７）受診相談窓口における相談件数」のグラフを掲載していることから、データが重複するため、6月15日をもって削除いたします。": "本グラフは、5月22日より「モニタリング指標（７）受診相談窓口における相談件数」のグラフを掲載していることから、データが重複するため、6月15日をもって削除いたします。",
-  "本グラフは、5月13日より「検査実施人数（陰性確認を除く）と陽性率の推移」（現在の「モニタリング指標(6)PCR検査の陽性率」）のグラフで検査実施人数を掲載していることから、データが重複するため、6月15日をもって削除いたします。": "本グラフは、5月13日より「検査実施人数（陰性確認を除く）と陽性率の推移」（現在の「モニタリング指標(6)PCR検査の陽性率」）のグラフで検査実施人数を掲載していることから、データが重複するため、6月15日をもって削除いたします。",
-  "本グラフは、5月8日より「検査実施件数」のグラフで健康安全研究センターが行った検査件数と医療機関等が行った件数の内訳を掲載していること、5月13日より「検査実施人数（陰性確認を除く）と陽性率の推移」（現在の「モニタリング指標(6)PCR検査の陽性率」）のグラフで検査実施人数を掲載していることから、データが重複するため、6月15日をもって削除いたします。": "本グラフは、5月8日より「検査実施件数」のグラフで健康安全研究センターが行った検査件数と医療機関等が行った件数の内訳を掲載していること、5月13日より「検査実施人数（陰性確認を除く）と陽性率の推移」（現在の「モニタリング指標(6)PCR検査の陽性率」）のグラフで検査実施人数を掲載していることから、データが重複するため、6月15日をもって削除いたします。",
-  "その他 参考指標": "その他 参考指標"
+  "（4）重症患者数 レベル1：100床 → レベル2：300床 → レベル3：700床": "（4）重症患者数 レベル1：100床 → レベル2：300床 → レベル3：700床"
 }
