+++ auto-i18n/result.csv
@@ -385,14 +385,3 @@ TAG_REMOVE,（4）重症患者数 レベル1 100床→レベル2 300床→レベ
 TAG_REMOVE,（5）入院患者数 レベル1 1,000床→レベル2 3,000床→レベル3 4,000床
 RUN,2020/06/01 07:55
 TAG_REMOVE,%{date}の全体累計
-RUN,2020/06/01 18:44
-TAG_ADD,{date}の全体累計
-RUN,2020/06/02 09:04
-TAG_ADD,検査実施人数（2月14日以前分を含む累計）
-TAG_ADD,検査実施人数 (日別)
-RUN,2020/06/02 09:07
-TAG_ADD,本グラフは、5月22日より「モニタリング指標（７）受診相談窓口における相談件数」のグラフを掲載していることから、データが重複するため、6月15日をもって削除いたします。
-TAG_ADD,本グラフは、5月13日より「検査実施人数（陰性確認を除く）と陽性率の推移」（現在の「モニタリング指標(6)PCR検査の陽性率」）のグラフで検査実施人数を掲載していることから、データが重複するため、6月15日をもって削除いたします。
-TAG_ADD,本グラフは、5月8日より「検査実施件数」のグラフで健康安全研究センターが行った検査件数と医療機関等が行った件数の内訳を掲載していること、5月13日より「検査実施人数（陰性確認を除く）と陽性率の推移」（現在の「モニタリング指標(6)PCR検査の陽性率」）のグラフで検査実施人数を掲載していることから、データが重複するため、6月15日をもって削除いたします。
-RUN,2020/06/02 09:10
-TAG_ADD,その他 参考指標
