+++ components/PositiveRateMixedChart.vue
@@ -316,31 +316,29 @@ const options: ThisTypedComponentOptionsWithRecordProps<
       ]
     },
     tableData() {
-      // 2月14日以前の分39件を累計値に足す
-      let cumulative = 39
       return this.labels
         .map((label, i) => {
-          let [dailySum, data] = [0, 0]
           return Object.assign(
             { text: dayjs(label).format('M/D') },
             ...this.tableHeaders.map((_, j) => {
-              switch (j) {
-                case 0: // 陽性者数
-                case 1: // 陰性者数
-                  dailySum += data = this.chartData[j][i]
-                  break
-                case 2: // 検査実施人数 (日別)
-                  cumulative += data = dailySum
-                  break
-                case 3: // 検査実施人数 (累計)
-                  data = cumulative
-                  break
-                case 4: // 陽性率
-                  data = this.chartData[2][i]
-                  break
+              if (j <= 1) {
+                const data = this.chartData[j]
+                return {
+                  [j]: data[i].toLocaleString()
+                }
               }
-              return {
-                [j]: data.toLocaleString()
+              if (j === 2) {
+                const positive = this.chartData[0]
+                const negative = this.chartData[1]
+                return {
+                  [j]: (positive[i] + negative[i]).toLocaleString()
+                }
+              }
+              if (j === 3) {
+                const data = this.chartData[2]
+                return {
+                  [j]: data[i].toLocaleString()
+                }
               }
             })
           )
