+++ components/TimeBarChart.vue
@@ -1,8 +1,5 @@
 <template>
   <data-view :title="title" :title-id="titleId" :date="date">
-    <template v-slot:attentionNote>
-      <slot name="attentionNote" />
-    </template>
     <template v-slot:description>
       <slot name="description" />
     </template>
