+++ components/cards/InspectionPersonsNumberCard.vue
@@ -8,15 +8,6 @@
       :date="data.date"
       :unit="$t('人')"
     >
-      <template v-slot:attentionNote>
-        <p>
-          {{
-            $t(
-              '本グラフは、5月13日より「検査実施人数（陰性確認を除く）と陽性率の推移」（現在の「モニタリング指標(6)PCR検査の陽性率」）のグラフで検査実施人数を掲載していることから、データが重複するため、6月15日をもって削除いたします。'
-            )
-          }}
-        </p>
-      </template>
       <template v-slot:description>
         <ul>
           <li>
