+++ components/cards/PositiveRateCard.vue
@@ -91,8 +91,7 @@ export default {
     const positiveRateTableLabels = [
       this.$t('陽性者数'),
       this.$t('陰性者数'),
-      this.$t('検査実施人数 (日別)'),
-      this.$t('検査実施人数（2月14日以前分を含む累計）'),
+      this.$t('検査実施人数'),
       this.$t('陽性率')
     ]
